requirements.txt
# LogoMaster Pro - Зависимости
# Основные библиотеки
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0

# Дополнительные библиотеки для расширенной функциональности
# tkinter-tooltip>=2.2.0  # Опционально для подсказок
# opencv-python>=4.8.0    # Опционально для расширенной обработки изображений

# Библиотеки для разработки и тестирования
# pytest>=7.4.0          # Для тестирования
# black>=23.0.0           # Для форматирования кода
# flake8>=6.0.0           # Для проверки стиля кода